requirements.txt
pandas
numpy
paho-mqtt
requests